+++ hacker-guides/encryption/encrypt-file/project.json
@@ -1,0 +1,12 @@
+{
+    "name": "encrypt-file-nodejs",
+    "root": "encrypt-file/nodejs",
+    "targets": {
+        "test-lit": {
+            "executor": "nx:run-commands",
+            "options": {
+                "command": "cd hacker-guides/encryption/encrypt-file && yarn && yarn test"
+            }
+        }
+    }
+}
