+++ hacker-guides/encryption/encrypt-file/src/index.ts
@@ -13,7 +13,7 @@ import { getEnv, getEthersSigner, getLitNodeClient } from "./utils";
 
 const ETHEREUM_PRIVATE_KEY = getEnv("ETHEREUM_PRIVATE_KEY");
 
-export const runExample = async () => {
+export const encryptFile = async () => {
   let litNodeClient: LitNodeClient;
 
   try {
