+++ hacker-guides/encryption/encrypt-file/test/index.spec.ts
@@ -2,7 +2,7 @@ import { expect } from "chai";
 import { promises as fs } from "fs";
 import { join } from "path";
 
-import { runExample } from "../src";
+import { encryptFile } from "../src";
 
 describe("Encrypting and decrypting a file", () => {
   afterEach(async () => {
@@ -24,7 +24,7 @@ describe("Encrypting and decrypting a file", () => {
     const originalPath = join(process.cwd(), "src", "loremIpsum.txt");
     const originalContent = await fs.readFile(originalPath, "utf8");
 
-    await runExample();
+    await encryptFile();
 
     const decryptedPath = join(
       process.cwd(),
