+++ hacker-guides/encryption/encrypt-large-file/project.json
@@ -1,0 +1,12 @@
+{
+    "name": "encrypt-large-file-nodejs",
+    "root": "encrypt-large-file/nodejs",
+    "targets": {
+        "test-lit": {
+            "executor": "nx:run-commands",
+            "options": {
+                "command": "cd hacker-guides/encryption/encrypt-large-file && yarn && yarn test"
+            }
+        }
+    }
+}
