+++ hacker-guides/encryption/encrypt-string/project.json
@@ -1,0 +1,12 @@
+{
+    "name": "encrypt-string-nodejs",
+    "root": "encrypt-string/nodejs",
+    "targets": {
+        "test-lit": {
+            "executor": "nx:run-commands",
+            "options": {
+                "command": "cd hacker-guides/encryption/encrypt-string && yarn && yarn test"
+            }
+        }
+    }
+}
