+++ hacker-guides/encryption/encrypt-string/test/index.spec.ts
@@ -1,10 +1,10 @@
 import { expect } from "chai";
 
-import { runExample } from "../src";
+import { encryptString } from "../src";
 
 describe("Encrypting and decrypting a string", () => {
   it("should encrypt and decrypt a string", async () => {
-    const decryptedString = await runExample();
+    const decryptedString = await encryptString();
     expect(decryptedString).to.equal(
       "The answer to life, the universe, and everything is 42."
     );
